# Send smaller number of requests in Apollo view change and preexec tests