We are are getting close to the maximum job time limit in Travis. I'm trying to optimise some of the tests to take less time.
This patch shows around 28 secs improvement.

The change is smaller number of ops started by the test.

Without the patch:
```
$ ctest -R skvbc_view_change_tests_v2merkle
Test project /home/tdimitrov/concord-bft/build
    Start 69: skvbc_view_change_tests_v2merkle
1/1 Test #69: skvbc_view_change_tests_v2merkle ...   Passed  332.40 sec

100% tests passed, 0 tests failed out of 1

Total Test time (real) = 332.59 sec
```

After the patch:
```
Test project /home/tdimitrov/concord-bft/build
    Start 69: skvbc_view_change_tests_v2merkle
1/1 Test #69: skvbc_view_change_tests_v2merkle ...   Passed  304.11 sec

100% tests passed, 0 tests failed out of 1

Total Test time (real) = 304.28 sec
```